# Refactored index.js to reduce bundle size by 30%