@Ben-Brady 2025-05-06
Also I think the `typeof arg !== "object"` check is unncessary, I didn't fail any tests when I removed it and I can't see why it would be neccessary.
However, I'm keeping it in currently as I don't want to accidentally break any code not covered by tests